requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.53
python-multipart==0.0.6
jinja2==3.1.2
qrcode[pil]==7.4.2
apscheduler==3.10.4
aiohttp==3.9.1
sse-starlette==1.6.5
python-dotenv==1.0.0